Changed location of interface types for Users and Tweets
